CTB-881 close bucket writers that fail to flush
It appears to be possible that a bucket writer can get it's self into a
state where it accepts appends but fails the flush. It's difficult to
pin down exactly the circumstances of what causes this, but it
seems reasonable to perform a close on the bucket writer on failure
in a similar way to the append method does.

Note the bucket writer is unable to implement this directly as
it is close calls flush so it'd end in a loop.